[Refactor] 알림 데이터 타입 변경

+++ backend/common/src/main/java/minionz/common/user_alarm/model/UserAlarm.java
@@ -17,7 +17,7 @@ public class UserAlarm extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long userAlarmId;
 
-    private Integer alarmStatus;
+    private Long alarmStatus;
     private String category;
 
     // UserAlarm : User = N : 1
